+++ src/tile-group/index.js
@@ -1,0 +1,11 @@
+import './style.scss';
+import { registerBlockType } from '@wordpress/blocks';
+import { InnerBlocks } from '@wordpress/block-editor';
+import metadata from './block.json';
+import Edit from './edit';
+
+registerBlockType( metadata.name, {
+	edit: Edit,
+	save: () => <InnerBlocks.Content />,
+	deprecated: [],
+} );